+++ app/components/mocks/index.js
@@ -50,4 +50,4 @@ const mocks = {
 };
 
 // switch this to mocks item to test [TODO: replace with storybooks]
-export default null;
+export default null; // MockChannelView;